Changed and fixed some typos/grammar.
There are a couple of places where either the code itself is not explained properly, e.g. why use `tasklogger`.  If the aim is to show the difference in efficiencies, then that must be said so explicitly.  I would rather that it be said explicitly here, than mentioning it in the lectures.  The notebooks should ideally be as complete as possible in themselves.