Changed for clarity.
I would suggest that once the notebook 05 in this folder is introduced, we should ask the participants to do the exercise with real data or even their own data so that they understand how the different DR algorithms will behave differently with the same data.  This point has been made in one of the notebooks, that needs to be really harped.